FTD_JupyterNotebook/requirements.txt
absl-py==1.4.0
alabaster==0.7.13
anyio==4.0.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.12.2
bleach==6.0.0
bokeh==2.4.3
Bottleneck==1.3.4
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==3.1.0
comm==0.1.4
contourpy==1.0.7
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docutils==0.18.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
fonttools==4.39.3
fqdn==1.5.1
idna==3.4
imagesize==1.4.1
importlib-metadata==6.1.0
ipykernel==6.19.2
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.0.3
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.1.0
jupyter-console==6.6.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter_core==5.3.0
jupyter-events==0.7.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.6
netrd==0.3.0
networkx==3.0
nibabel==5.0.1
nilearn==0.10.0
notebook==7.0.3
notebook_shim==0.2.3
numexpr==2.8.1
numpy==1.22.3
numpydoc==1.2.1
olefile==0.46
openpyxl==3.0.9
ortools==9.6.2534
overrides==7.4.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
panel==0.14.4
param==1.13.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==21.2.4
platformdirs==3.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyct==0.5.0
Pygments==2.14.0
pyparsing==3.0.4
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.1
pyviz-comms==2.2.1
PyYAML==6.0
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.2
scikit-learn==1.2.2
scipy==1.11.1
seaborn==0.12.2
Send2Trash==1.8.2
setuptools==61.2.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==2.0.1
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.2
statsmodels==0.13.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
unicodedata2==15.0.0
uri-template==1.3.0
urllib3==1.26.15
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
wheel==0.37.1
widgetsnbextension==4.0.8
xlrd==2.0.1
zipp==3.15.0
